+++ lifetimes/fitters/pareto_nbd_fitter.py
@@ -95,7 +95,7 @@ class ParetoNBDFitter(BaseFitter):
             with additional properties like params_ and methods like predict
 
         """
-        frequency = asarray(frequency)
+        frequency = asarray(frequency).astype(int)
         recency = asarray(recency)
         T = asarray(T)
